fastapi_backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
motor==3.3.2
pymongo==4.6.0
google-generativeai==0.8.0
redis==5.0.1
slowapi==0.1.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pytest==7.4.3
pytest-asyncio==0.21.1